Moved the whole auth login into a separate app
