+++ docker-app/qfieldcloud/core/serializers.py
@@ -1,8 +1,6 @@
-from django.conf import settings
-from django.contrib.auth import authenticate, get_user_model
-from django.utils.translation import gettext as _
+from django.contrib.auth import get_user_model
+from qfieldcloud.authentication.models import AuthToken
 from qfieldcloud.core.models import (
-    AuthToken,
     Delta,
     ExportJob,
     Job,
@@ -68,103 +66,6 @@ class ProjectSerializer(serializers.ModelSerializer):
             "user_role_origin",
         )
         model = Project
-
-
-class LoginSerializer(serializers.Serializer):
-    """Based on https://github.com/Tivix/django-rest-auth/blob/rest_auth/serializers.py#L19"""
-
-    username = serializers.CharField(required=False, allow_blank=True)
-    email = serializers.EmailField(required=False, allow_blank=True)
-    password = serializers.CharField(style={"input_type": "password"})
-
-    def authenticate(self, **kwargs):
-        return authenticate(self.context["request"], **kwargs)
-
-    def _validate_email(self, email, password):
-        user = None
-
-        if email and password:
-            user = self.authenticate(email=email, password=password)
-        else:
-            msg = _('Must include "email" and "password".')
-            raise ValidationError(msg)
-
-        return user
-
-    def _validate_username(self, username, password):
-        user = None
-
-        if username and password:
-            user = self.authenticate(username=username, password=password)
-        else:
-            msg = _('Must include "username" and "password".')
-            raise ValidationError(msg)
-
-        return user
-
-    def _validate_username_email(self, username, email, password):
-        user = None
-
-        if email and password:
-            user = self.authenticate(email=email, password=password)
-        elif username and password:
-            user = self.authenticate(username=username, password=password)
-        else:
-            msg = _('Must include either "username" or "email" and "password".')
-            raise ValidationError(msg)
-
-        return user
-
-    def validate(self, attrs):
-        username = attrs.get("username")
-        email = attrs.get("email")
-        password = attrs.get("password")
-
-        user = None
-
-        if "allauth" in settings.INSTALLED_APPS:
-            from allauth.account import app_settings
-
-            # Authentication through email
-            if (
-                app_settings.AUTHENTICATION_METHOD
-                == app_settings.AuthenticationMethod.EMAIL
-            ):
-                user = self._validate_email(email, password)
-
-            # Authentication through username
-            elif (
-                app_settings.AUTHENTICATION_METHOD
-                == app_settings.AuthenticationMethod.USERNAME
-            ):
-                user = self._validate_username(username, password)
-
-            # Authentication through either username or email
-            else:
-                user = self._validate_username_email(username, email, password)
-
-        else:
-            # Authentication without using allauth
-            if email:
-                try:
-                    username = User.objects.get(email__iexact=email).get_username()
-                except User.DoesNotExist:
-                    pass
-
-            if username:
-                user = self._validate_username_email(username, "", password)
-
-        # Did we get back an active user?
-        if user:
-            if not user.is_active:
-                msg = _("User account is disabled.")
-                raise ValidationError(msg)
-        else:
-            msg = _("Unable to log in with provided credentials.")
-            raise ValidationError(msg)
-
-        attrs["user"] = user
-        return attrs
 
 
 class CompleteUserSerializer(serializers.ModelSerializer):
