+++ docker-app/qfieldcloud/core/tests/test_permission.py
@@ -1,12 +1,7 @@
 import logging
 
-from qfieldcloud.core.models import (
-    AuthToken,
-    Organization,
-    Project,
-    ProjectCollaborator,
-    User,
-)
+from qfieldcloud.authentication.models import AuthToken
+from qfieldcloud.core.models import Organization, Project, ProjectCollaborator, User
 from rest_framework import status
 from rest_framework.test import APITestCase
