+++ docker-app/qfieldcloud/core/tests/test_project.py
@@ -1,6 +1,7 @@
 import logging
 
-from qfieldcloud.core.models import AuthToken, Project, ProjectCollaborator, User
+from qfieldcloud.authentication.models import AuthToken
+from qfieldcloud.core.models import Project, ProjectCollaborator, User
 from rest_framework import status
 from rest_framework.test import APITestCase
