+++ docker-app/qfieldcloud/core/tests/test_qfield_file.py
@@ -6,8 +6,9 @@ import time
 import psycopg2
 import requests
 from django.http.response import HttpResponseRedirect
+from qfieldcloud.authentication.models import AuthToken
 from qfieldcloud.core.geodb_utils import delete_db_and_role
-from qfieldcloud.core.models import AuthToken, Geodb, Project, User
+from qfieldcloud.core.models import Geodb, Project, User
 from rest_framework import status
 from rest_framework.test import APITransactionTestCase
