+++ docker-app/qfieldcloud/core/tests/test_queryset.py
@@ -1,8 +1,8 @@
 import logging
 
+from qfieldcloud.authentication.models import AuthToken
 from qfieldcloud.core import querysets_utils
 from qfieldcloud.core.models import (
-    AuthToken,
     Organization,
     OrganizationMember,
     Project,
